Rename the extend area.
