+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/Texting/ManageSwitchSpecialActionSection.cs
@@ -7,7 +7,7 @@ using osu.Framework.Allocation;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Manage
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Texting
 {
     public class ManageSwitchSpecialActionSection : SpecialActionSection<ManageEditModeSpecialAction>
     {
